+++ cpp/include/cuspatial/trajectory.hpp
@@ -21,7 +21,6 @@
 #include <rmm/mr/device/default_memory_resource.hpp>
 
 namespace cuspatial {
-namespace experimental {
 
 /**
  * @brief Derive trajectories from object ids, points, and timestamps.
@@ -120,5 +119,4 @@ std::unique_ptr<cudf::table> trajectory_bounding_boxes(
   cudf::column_view const& y,
   rmm::mr::device_memory_resource* mr = rmm::mr::get_default_resource());
 
-}  // namespace experimental
 }  // namespace cuspatial